refactor: rename authorization annotations for consistency
