+++ src/main/java/com/example/newsper/annotation/AdminOnly.java
@@ -1,4 +1,4 @@
-package com.example.newsper.annotations;
+package com.example.newsper.annotation;
 
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -10,8 +10,6 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@PreAuthorize("isAuthenticated() && !hasAuthority(T(com.example.newsper.constant.UserRole).ASSOCIATE.toString())")
+@PreAuthorize("hasAuthority(T(com.example.newsper.constant.UserRole).ADMIN.toString())")
 @SecurityRequirement(name = "Authorization")
-// not associate user(regular member) only
-public @interface MemberOnly {
-}
+public @interface AdminOnly {}
