+++ src/main/java/com/example/newsper/annotation/Unauthorized.java
@@ -1,6 +1,5 @@
-package com.example.newsper.annotations;
+package com.example.newsper.annotation;
 
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.security.access.prepost.PreAuthorize;
 
 import java.lang.annotation.ElementType;
@@ -10,6 +9,6 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@PreAuthorize("hasAuthority(T(com.example.newsper.constant.UserRole).ADMIN.toString())")
-@SecurityRequirement(name = "Authorization")
-public @interface AdminOnly {}
+@PreAuthorize("!isAuthenticated()")
+public @interface Unauthorized {
+}
