--- src/main/java/com/example/newsper/annotations/MustNotAuthorized.java
@@ -1,14 +1,0 @@
-package com.example.newsper.annotations;
-
-import org.springframework.security.access.prepost.PreAuthorize;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@PreAuthorize("!isAuthenticated()")
-public @interface MustNotAuthorized {
-}
