+++ src/main/java/com/example/newsper/api/CommentApiController.java
@@ -1,6 +1,6 @@
 package com.example.newsper.api;
 
-import com.example.newsper.annotations.MustAuthorized;
+import com.example.newsper.annotation.Authorized;
 import com.example.newsper.constant.ErrorCode;
 import com.example.newsper.dto.AddCommentDto;
 import com.example.newsper.dto.CommentDto;
@@ -69,7 +69,7 @@ public class CommentApiController {
     }
 
     @PostMapping("/comment")
-    @MustAuthorized
+    @Authorized
     @Operation(summary = "댓글 작성", description = "게시글에 댓글을 작성합니다. 액세스 토큰 필요.")
     @ApiResponse(responseCode = "201", description = "성공")
     public ResponseEntity<?> create(
@@ -86,7 +86,7 @@ public class CommentApiController {
     }
 
     @PatchMapping("/comment/{id}")
-    @MustAuthorized
+    @Authorized
     @Operation(summary = "댓글 수정", description = "댓글을 수정합니다.")
     @ApiResponse(responseCode = "200", description = "성공")
     public ResponseEntity<?> update(
@@ -105,7 +105,7 @@ public class CommentApiController {
     }
 
     @DeleteMapping("/comment/{id}")
-    @MustAuthorized
+    @Authorized
     @Operation(summary = "댓글 삭제", description = "댓글을 삭제합니다.")
     @ApiResponse(responseCode = "200", description = "성공")
     public ResponseEntity<?> delete(
